Lias version of the analysis script - messy
This is Lia Washington's version of the analysis script. It is based on (and builds upon) the original version of the script written by MKS, but it is extremely discombobulated as I keep needing to do new things. Since the MKS version is the preregistered version and the only file housing the analysis plan, I am going to restore it to its original form and not touch it again for fear of changing it too many times and losing the plot