Fixed Bugs
Issue: After the 1st trial, the study would just pause on the final frame of the video (red curtain) and did not proceed.
Solution: Originally the audio and the video were seperate, so the code had a requiredAudioCount. Since we are no longer using audio, we could not reach that requiredAudioCount!

Also updated the code for a few small details and updated the actualTrial for loop to accomodate 4 trials instead of 6.