Re-created my version of the v1 analysis script
This version of the v1 script is the "updated" version that runs with the data as it is. I wanted to preserve the original script MKS wrote, so I had this file but later converted it into v2_analysis script. However neither the original v1 analysis nor the v2 analysis actually runs for the v1 data due to naming and formatting choices I made throughout v1. Hence, the need for this file :)